requirements.txt
# Core bot dependencies
aiogram==3.20.0
aiohttp==3.10.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
apscheduler==3.10.4

# Redis for caching and queues (Phase 1)
redis==5.2.1
hiredis==3.1.0

# PostgreSQL dependencies (Phase 2-3)
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0

# RQ dependencies (Phase 5)
rq==1.16.1
rq-dashboard==0.8.5
rq-scheduler==0.13.1